extracted reusable part of review into Summary component
